RestfulAPI/requirements.txt
aniso8601==9.0.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
comtypes==1.1.11
Deprecated==1.2.13
Django==4.0.4
django-storages==1.12.3
docopt==0.6.2
# Editable install with no version control (drqa==0.1.0)
-e d:\language\python\virtual assistant\drqa
elastic-transport==8.1.0
elasticsearch==8.1.1
Flask==2.1.2
Flask-Cors==3.0.9
Flask-Helper==1.2.7
flask-marshmallow==0.14.0
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.20.9
greenlet==1.1.2
gunicorn==19.9.0
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.71
jsonschema==4.5.1
MarkupSafe==2.1.1
marshmallow==3.15.0
mysql-connector-python==8.0.29
nltk==3.7
nose==1.3.7
numpy==1.22.3
packaging==21.3
pexpect==4.2.1
pipwin==0.5.2
prettytable==3.2.0
protobuf==3.20.1
ptyprocess==0.7.0
PyAudio @ file:///C:/Users/84337/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pycodestyle==2.8.0
pydub==0.25.1
pyjsparser==2.7.1
pyparsing==3.0.7
pypiwin32==223
PyPrind==2.11.3
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dotenv==0.20.0
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==6.0
redis==4.3.1
regex==2022.3.15
requests==2.21.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.36
sqlparse==0.4.2
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.11.0
tqdm==4.63.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.24.3
uuid==1.30
validate-email==1.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1